Return ranges to registering nodes
ClusterStateResponse used to be just a string of node addresses.
Now it is a struct called ClusterMeta containing additional meta
data, specifically the ranges for which the registering node is
responsible and their current owners.

This allows the registering node to request it's dataset from
the rest of the cluster.